refactor : layout - Header
디자인 수정
리팩토링(파일, css)

Related to : #134
