+++ src/app/cart/page.tsx
@@ -22,7 +22,7 @@ export default function cartPage() {
 		cart.map((item) => item.isbn),
 	);
 	// useUser에서 호출한 로그인 상태
-	const user = useUser();
+	const { isLoggedIn } = useUser();
 
 	// 전체 선택 체크박스 state
 	const [selectAll, setSelectAll] = useState<boolean>(true);
@@ -119,7 +119,7 @@ export default function cartPage() {
 	// 주문하기 버튼 클릭 시 동작하는 함수
 	const handleOrderButtonClick = () => {
 		// 로그인 상태가 아닐 경우
-		if (!user) {
+		if (!isLoggedIn) {
 			if (confirm('로그인이 필요한 서비스입니다 로그인 하시겠습니까?')) {
 				// 로그인 페이지로 이동
 				router.push('/login');
