+++ src/hooks/useUser.ts
@@ -1,22 +1,27 @@
+'use client';
+
 import { getUser } from '@/apis/community/getUser';
 import { useEffect, useState } from 'react';
 
 export const useUser = () => {
 	// 유저 state
-	const [user, setUser] = useState<string | null>(null);
+	const [isLoggedIn, setIsLoggedIn] = useState<string | null>(null);
+	// 유저 name state
+	const [userName, setUserName] = useState<string | null>(null);
 
 	// getUser api에서 user_id와 error를 뿌려줄 useEffect
 	useEffect(() => {
 		const fetchUser = async () => {
 			try {
-				const { user_id } = await getUser();
-				setUser(user_id as string);
+				const { user_id, user_name } = await getUser();
+				setIsLoggedIn(user_id as string);
+				setUserName(user_name);
 			} catch (error) {
-				setUser(null);
+				setIsLoggedIn(null);
 			}
 		};
 		fetchUser();
 	}, []);
 
-	return user;
+	return { isLoggedIn, userName };
 };
