+++ src/styles/layout/header/bottomWrapper/search/recentSearch/lastestWord.module.css
