+++ src/styles/layout/header/bottomWrapper/search/recentSearch/recentSearch.module.css
