+++ src/styles/layout/header/bottomWrapper/search/searchResult.module.css
