+++ src/styles/layout/header/header.module.css
@@ -1,7 +1,6 @@
 /* 헤더 틀 */
 .container {
-	/* width: 100vw; */
-	min-height: 140px;
+	height: 170px;
 
 	display: flex;
 	justify-content: center;
@@ -14,49 +13,4 @@
 
 	display: flex;
 	flex-direction: column;
-}
-
-/* 헤더 상단 */
-.topWrapper {
-	/* min-height: 60px; */
-	min-height: 80px;
-
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-
-	/* border: 3px solid black; */
-}
-.logoWrapper {
-	height: 80px;
-
-	display: flex;
-	align-items: center;
-}
-.inputWrapper {
-	width: 400px;
-	height: 80px;
-}
-.loginWrapper {
-	width: inherit;
-	height: 30px;
-	gap: 12px;
-	font-size: 12px;
-
-	display: flex;
-	align-items: center;
-	justify-content: end;
-
-	cursor: pointer;
-}
-.loginWrapper > :first-child {
-	color: #6892e8;
-}
-
-/* 헤더 하단 */
-.bottomWrapper {
-	min-height: 60px;
-
-	display: flex;
-	justify-content: space-between;
 }
